Use integers with well defined bit widths for ProgramInfo data. This requires a plug-in recompile.
git-svn-id: http://svn.mythtv.org/svn/trunk@22999 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/libs/libmyth/programinfoupdater.cpp
@@ -8,7 +8,7 @@ uint qHash(const PIKey &k)
 
 void ProgramInfoUpdater::insert(
     uint     chanid, const QDateTime &recstartts,
-    PIAction action, long long        filesize)
+    PIAction action, uint64_t         filesize)
 {
     QMutexLocker locker(&lock);
     if (kPIUpdate == action || kPIUpdateFileSize == action)
